+++ RELEASE_NOTES.md
@@ -6,6 +6,8 @@ Changes in version 0.6.18
   - Allow child_scopes() and set_child_scopes() methods more time to
     execute as they read from and write to disk.
   - Added ChildScopes.Timeout configuration parameter to Zmq.ini
+  - Added API for activation of in-card result actions - see
+    ScopeBase::activate_result_action() and ActivationResponse::Status::UpdateResult.
 
 Changes in version 0.6.17
 =========================